+++ areas/functionapp/src/AzureMcp.FunctionApp/Commands/FunctionAppJsonContext.cs
@@ -8,6 +8,7 @@ using AzureMcp.FunctionApp.Models;
 namespace AzureMcp.FunctionApp.Commands;
 
 [JsonSerializable(typeof(FunctionAppListCommand.FunctionAppListCommandResult))]
+[JsonSerializable(typeof(FunctionAppGetCommand.FunctionAppGetCommandResult))]
 [JsonSerializable(typeof(FunctionAppInfo))]
 [JsonSourceGenerationOptions(PropertyNamingPolicy = JsonKnownNamingPolicy.CamelCase)]
 internal partial class FunctionAppJsonContext : JsonSerializerContext;