+++ areas/functionapp/src/AzureMcp.FunctionApp/Options/FunctionApp/FunctionAppListOptions.cs
@@ -1,8 +1,6 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-using AzureMcp.Core.Options;
-
-namespace AzureMcp.FunctionApp.Options;
+namespace AzureMcp.FunctionApp.Options.FunctionApp;
 
 public class FunctionAppListOptions : SubscriptionOptions;